--- a/src/methods/area.jl
+++ b/src/methods/area.jl
@@ -123,12 +123,22 @@ function _signed_area(::Type{T}, geom) where T
     # Close curve, even if last point isn't explicitly repeated
     np = GI.npoint(geom)
     np == 0 && return area
+    first = true
+    local pfirst, p1
     # Integrate the area under the curve
-    p1 = GI.getpoint(geom, np)
     for p2 in GI.getpoint(geom)
+        # Skip the first and do it later 
+        # This lets us work withing one iteration, which means on C call when 
+        # using points from external libraries
+        if first
+            p1 = pfirst = p2
+            first = false
+            continue
+        end
         # Accumulate the area into `area`
         area += GI.x(p1) * GI.y(p2) - GI.y(p1) * GI.x(p2)
-        p1 = p2
     end
+    p2 = pfirst
+    area += GI.x(p1) * GI.y(p2) - GI.y(p1) * GI.x(p2)
     return T(area / 2)
 end