--- a/src/methods/area.jl
+++ b/src/methods/area.jl
@@ -120,16 +120,16 @@ to be closed.
 =#
 function _signed_area(::Type{T}, geom) where T
     area = zero(T)
-    # Close curve, even if last point isn't explicitly repeated
     np = GI.npoint(geom)
     np == 0 && return area
+
     first = true
     local pfirst, p1
     # Integrate the area under the curve
     for p2 in GI.getpoint(geom)
         # Skip the first and do it later 
-        # This lets us work withing one iteration, which means on C call when 
-        # using points from external libraries
+        # This lets us work within one iteration over geom, 
+        # which means on C call when using points from external libraries.
         if first
             p1 = pfirst = p2
             first = false
@@ -137,7 +137,10 @@ function _signed_area(::Type{T}, geom) where T
         end
         # Accumulate the area into `area`
         area += GI.x(p1) * GI.y(p2) - GI.y(p1) * GI.x(p2)
+        p1 = p2
     end
+    # Complete the last edge.
+    # If the first and last where the same this will be zero
     p2 = pfirst
     area += GI.x(p1) * GI.y(p2) - GI.y(p1) * GI.x(p2)
     return T(area / 2)