With all the fixes in LibGEOS this gets us to ~20x slower territory for #32 

```julia
julia> @benchmark LG.area($pLG)
BenchmarkTools.Trial: 10000 samples with 990 evaluations.
 Range (min … max):  32.888 ns … 71.911 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     33.754 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   35.655 ns ±  4.169 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆▄█▇ ▄▁▁▄  ▂             ▁    ▂ ▁▄▂   ▃    ▁                ▁
  █████████████▇▇█▄▇█▅▅▆▇▅▇█▅▄▃▆█▆███▆▆██▇▅▆▄█▇▄▅▆▆▆▅▅▅▇▅▆▅▄▅ █
  32.9 ns      Histogram: log(frequency) by time      49.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark GO.area($pLG)
BenchmarkTools.Trial: 10000 samples with 143 evaluations.
 Range (min … max):  714.762 ns …  5.840 ms  ┊ GC (min … max): 0.00% … 5.31%
 Time  (median):       1.111 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):     2.401 μs ± 73.988 μs  ┊ GC (mean ± σ):  2.36% ± 0.08%

    ▁█ ▁  ▁      ▃▂                                             
  ▁▂██▅█▆▅█▅▄▄▅▆▆██▃▃▂▂▁▂▂▃▃▃▃▃▃▃▃▃▂▂▂▂▂▂▁▁▂▂▂▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁ ▂
  715 ns          Histogram: frequency by time         2.45 μs <

 Memory estimate: 80 bytes, allocs estimate: 3.

julia> @benchmark GO.area($pGI)
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  18.195 ns … 461.430 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     18.467 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   22.606 ns ±  12.344 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █   ▁ ▁▂   ▃ ▃ ▅ ▅  ▁ ▂  ▁                       ▁▁▁▁▂▂▂▁▁   ▂
  █▇█▁██████▆█▆█████▅▇█▆█▇▆███▄▅▅▆▆▆▅▆▆▆▇▇▆▆▇███▇█████████████ █
  18.2 ns       Histogram: log(frequency) by time      37.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

(and 2x faster than libgeos when working with native geoms!)


The idea with deleting all the handling of closed/open polygons is that the last point will contribute zero area anyway if its a repeat of the first one - so we can just ignore it.