@skygering 2024-01-02
Looks great! Dang, the LibGEOS performance is still painful! But excited about the improvements with native polygons!