@rafaqz 2024-01-02
I got it down to 3x faster than that with other changes...